requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.19.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.17.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
magic-filter==1.0.12
multidict==6.6.4
netaddr==1.3.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.0.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.43
structlog==24.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
yarl==1.20.1
yookassa==3.7.1
